+++ src/pages/Homepage/Homepage.tsx
@@ -1,28 +1,10 @@
-import React, { useState } from "react";
-import { Grid, useDisclosure } from "@chakra-ui/react";
-
-import { MintingForm } from "components/MintingForm";
-import { MintingModal } from "components/MintingModal";
-
-export type MintingStatusType = {
-  success: boolean;
-  status: string;
-};
+import React from "react";
+import { Grid, Text } from "@chakra-ui/react";
 
 export const Homepage = () => {
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const [mintingStatus, setMintingStatus] = useState<MintingStatusType | null>(
-    null
-  );
-
   return (
     <Grid minH="100vh" p={3}>
-      <MintingForm onOpen={onOpen} setMintingStatus={setMintingStatus} />
-      <MintingModal
-        isOpen={isOpen}
-        onClose={onClose}
-        mintingStatus={mintingStatus}
-      />
+      <Text size="3xl">Home Page!</Text>
     </Grid>
   );
 };
